updated files
changed Profile dialog to routing- it worked. The issue was the routing bug that is the correct naming: the singleMovieCard it now reading in the console but still not reendering any content in the browser. But the dialogs for Genre and Director are working now, but still not showing any content